Populate sltl::glsl::output_glsl class
Move the majority of glsl specific code from output.h/cpp to output_glsl.h/cpp. Unit tests updated to use sltl::glsl::output_glsl type.
