+++ sltl/src/output/output.cpp
@@ -38,49 +38,6 @@ namespace
     }
   }
 
-  bool is_variable_omitted(const ns::syntax::variable_declaration& vd, ns::output::layout_manager& layout_manager)
-  {
-    if(vd._qualifier != ns::core::qualifier_storage::none)
-    {
-      bool is_omitted = false;
-
-      if(vd._semantic == ns::core::semantic::none)
-      {
-        auto& layout_map = layout_manager.get_layout_map(vd);
-
-        if(layout_map.is_layout_enabled())
-        {
-          layout_map.set_location_next(layout_map.get_location_next(vd));
-        }
-
-        is_omitted = true;
-      }
-      else if(ns::glsl::is_variable_built_in(vd))
-      {
-        is_omitted = true;
-      }
-
-      return is_omitted;
-    }
-
-    return false;
-  }
-
-  bool is_layout_flag_valid(const ns::syntax::variable_declaration& vd, ns::output::layout_flags flag)
-  {
-    switch(vd._qualifier)
-    {
-      case ns::core::qualifier_storage::in:
-        return (flag == ns::output::layout_flags::flag_in);
-      case ns::core::qualifier_storage::out:
-        return (flag == ns::output::layout_flags::flag_out);
-      case ns::core::qualifier_storage::uniform:
-        return (flag == ns::output::layout_flags::flag_uniform);
-      default:
-        return false;
-    }
-  }
-
   template<typename T>
   auto to_string(T t) -> typename std::enable_if<std::is_floating_point<T>::value, std::wstring>::type
   {
@@ -181,24 +138,6 @@ namespace
     return nullptr;
   }
 
-  const wchar_t* to_version_string(ns::output_version version)
-  {
-    switch(version)
-    {
-      case ns::output_version::v330:
-        return L"#version 330";
-      default:
-        assert(version == ns::output_version::none);
-    }
-
-    return nullptr;
-  }
-
-  std::wstring get_type_name(const ns::language::type& type, ns::detail::enum_flags<ns::output_flags> flags)
-  {
-    return ns::glsl::to_type_string(flags.has_flag<ns::output_flags::flag_transpose_type>() ? type.transpose() : type);
-  }
-
   std::wstring get_parameter_name(const ns::syntax::parameter_declaration& pd)
   {
     std::wstringstream ss(ns::glsl::to_parameter_prefix_string(pd._qualifier), std::ios::in | std::ios::out | std::ios::ate);
@@ -208,39 +147,6 @@ namespace
     ss << pd._name;
 
     return ss.str();
-  }
-
-  std::wstring get_qualifier_layout(const ns::syntax::variable_declaration& vd, ns::output::layout_manager& layout_manager)
-  {
-    std::wstringstream ss;
-
-    if(vd._semantic != ns::core::semantic::none)
-    {
-      auto& layout_map = layout_manager.get_layout_map(vd);
-
-      if(layout_map.is_layout_enabled())
-      {
-        const auto result = layout_map.insert(vd);
-
-        ss << L"layout(location = ";
-        ss << result.first;
-        ss << L')';
-
-        assert(result.second);
-      }
-    }
-
-    return ss.str();
-  }
-
-  const wchar_t* get_qualifier_storage(const ns::syntax::variable_declaration& vd)
-  {
-    if(ns::glsl::is_variable_built_in(vd))
-    {
-      throw std::exception();//TODO: exception type and message
-    }
-
-    return ns::glsl::to_qualifier_string(vd._qualifier);
   }
 
   std::wstring get_zero_initialization(const ns::language::type& type)
@@ -272,44 +178,10 @@ namespace
 
     return value;
   }
-
-  ns::detail::enum_flags<ns::output::layout_flags> get_default_layout_flags(ns::core::shader_stage stage)
-  {
-    ns::detail::enum_flags<ns::output::layout_flags> flags = ns::output::layout_flags::flag_none;
-
-    switch(stage)
-    {
-      case ns::core::shader_stage::vertex:
-        flags = ns::output::layout_flags::flag_out;
-        break;
-      case ns::core::shader_stage::geometry:
-        flags = ns::output::layout_flags::flag_out |
-                ns::output::layout_flags::flag_in;
-        break;
-      case ns::core::shader_stage::fragment:
-        flags = ns::output::layout_flags::flag_in;
-        break;
-      default:
-        assert(stage == ns::core::shader_stage::test);
-        break;
-    }
-
-    return flags;
-  }
 }
 
-// output definitions
-
-ns::output::output(core::shader_stage stage, output_version version, detail::enum_flags<output_flags> flags) : output(stage, layout_manager(get_default_layout_flags(stage)), version, flags)
+ns::output::output(core::shader_stage stage, detail::enum_flags<output_flags> flags) : _flags(flags), _indent_count(0U), _stage(stage)
 {
-}
-
-ns::output::output(core::shader_stage stage, layout_manager&& manager, output_version version, detail::enum_flags<output_flags> flags) : _indent_count(0), _flags(flags), _stage(stage), _layout_manager(std::move(manager))
-{
-  if(const auto v_str = to_version_string(version))
-  {
-    _ss << v_str << std::endl;
-  }
 }
 
 std::wstring ns::output::get_result() const
@@ -344,59 +216,9 @@ ns::syntax::action_return_t ns::output::operator()(const syntax::io_block&, bool
                     ns::syntax::action_return_t::step_out;
 }
 
-ns::syntax::action_return_t ns::output::operator()(const syntax::variable_declaration& vd, bool is_start)
-{
-  ns::syntax::action_return_t return_val;
-
-  if(is_start)
-  {
-    if(!is_variable_omitted(vd, _layout_manager))
-    {
-      line_begin();
-
-      const auto qualifier_layout = get_qualifier_layout(vd, _layout_manager);
-      const auto qualifier_storage = get_qualifier_storage(vd);
-
-      if(!qualifier_layout.empty())
-      {
-        _ss << qualifier_layout << L' ';
-      }
-
-      if(qualifier_storage)
-      {
-        _ss << qualifier_storage << L' ';
-      }
-
-      _ss << get_type_name(vd.get_type(), _flags) << L' ' << glsl::get_variable_name(vd);
-
-      if(vd.has_initializer())
-      {
-        _ss << L" = ";
-      }
-
-      return_val = ns::syntax::action_return_t::step_in;
-    }
-    else
-    {
-      return_val = ns::syntax::action_return_t::step_over;
-    }
-  }
-  else
-  {
-    if(!is_variable_omitted(vd, _layout_manager))
-    {
-      line_end();
-    }
-
-    return_val = ns::syntax::action_return_t::step_out;
-  }
-
-  return return_val;
-}
-
 ns::syntax::action_return_t ns::output::operator()(const syntax::parameter_declaration& pd)
 {
-  _ss << get_type_name(pd.get_type(), _flags) << L' ' << get_parameter_name(pd);
+  _ss << get_type_name(pd.get_type()) << L' ' << get_parameter_name(pd);
 
   return ns::syntax::action_return_t::step_out;
 }
@@ -456,7 +278,7 @@ ns::syntax::action_return_t ns::output::operator()(const syntax::temporary& t, b
 
     if(!initializer || !sltl::is_type<syntax::constructor_call>(initializer))
     {
-      _ss << get_type_name(type, _flags) << L'(';
+      _ss << get_type_name(type) << L'(';
     }
 
     if(!initializer)
@@ -671,7 +493,7 @@ ns::syntax::action_return_t ns::output::operator()(const syntax::constructor_cal
 {
   if(is_start)
   {
-    _ss << get_type_name(cc.get_type(), _flags) << L'(';
+    _ss << get_type_name(cc.get_type()) << L'(';
   }
   else
   {
@@ -762,7 +584,7 @@ ns::syntax::action_return_t ns::output::operator()(const syntax::function_defini
   if(is_start)
   {
     line_begin();
-    _ss << get_type_name(fd.get_type(), _flags) << L' ' << fd._name << L'(';
+    _ss << get_type_name(fd.get_type()) << L' ' << fd._name << L'(';
   }
   else
   {
@@ -863,171 +685,4 @@ void ns::output::line_end(bool has_semi_colon)
   }
 
   _ss << std::endl;
-}
-
-// output::layout_map_key definitions
-
-ns::output::layout_map_key::layout_map_key(const ns::syntax::variable_declaration& vd) : _s(vd._semantic), _idx(vd._semantic_index)
-{
-  assert(vd._qualifier == ns::core::qualifier_storage::in ||
-         vd._qualifier == ns::core::qualifier_storage::out ||
-         vd._qualifier == ns::core::qualifier_storage::uniform);
-}
-
-ns::output::layout_map_key::layout_map_key(const layout_map_key& key) : _s(key._s), _idx(key._idx)
-{
-}
-
-bool ns::output::layout_map_key::operator<(const ns::output::layout_map_key& rhs) const
-{
-  return ((_s < rhs._s) || (!(rhs._s < _s) && (_idx < rhs._idx)));
-}
-
-// output::layout_map definitions
-
-ns::output::layout_map::layout_map(layout_flags flag) : _flag(flag), _location_next(0U)
-{
-}
-
-ns::output::layout_map::layout_map(layout_map&& map) : _flag(map._flag), _location_next(map._location_next), _location_map(std::move(map._location_map))
-{
-}
-
-std::pair<ns::output::layout_index_t, bool> ns::output::layout_map::insert(const syntax::variable_declaration& vd)
-{
-  // Ensure the storage qualifier matches the map's flag and that the semantic isn't 'none'
-  if(!is_layout_flag_valid(vd, _flag) || (vd._semantic == core::semantic::none))
-  {
-    throw std::exception();//TODO: exception type and message
-  }
-
-  const auto result = _location_map.emplace(layout_map_key(vd), _location_next);
-
-  // If an insertion occurred then increase the location index by an
-  // increment that is determined using the variable_declaration's type
-  if(result.second)
-  {
-    _location_next = get_location_next(vd);
-  }
-
-  return std::make_pair(result.first->second, result.second);
-}
-
-bool ns::output::layout_map::is_layout_enabled() const
-{
-  return (_flag != layout_flags::flag_none);
-}
-
-bool ns::output::layout_map::is_layout_qualified(const syntax::variable_declaration& vd) const
-{
-  // Ensure the storage qualifier matches the map's flag and that the semantic isn't 'none'
-  if(!is_layout_flag_valid(vd, _flag) || (vd._semantic == core::semantic::none))
-  {
-    throw std::exception();//TODO: exception type and message
-  }
-
-  return (_location_map.find(layout_map_key(vd)) != _location_map.end());
-}
-
-void ns::output::layout_map::set_location_next(layout_index_t location)
-{
-  _location_next = location;
-}
-
-ns::output::layout_index_t ns::output::layout_map::get_location(const syntax::variable_declaration& vd) const
-{
-  // Ensure the storage qualifier matches the map's flag and that the semantic isn't 'none'
-  if(!is_layout_flag_valid(vd, _flag) || (vd._semantic == core::semantic::none))
-  {
-    throw std::exception();//TODO: exception type and message
-  }
-
-  // Throws std::out_of_range if the key doesn't exist in the map
-  return _location_map.at(layout_map_key(vd));
-}
-
-ns::output::layout_index_t ns::output::layout_map::get_location_next(const syntax::variable_declaration& vd) const
-{
-  // Ensure the storage qualifier matches the map's flag
-  if(!is_layout_flag_valid(vd, _flag))
-  {
-    throw std::exception();//TODO: exception type and message
-  }
-
-  // Types: scalar, vector
-  // A single location index, or slot, is consumed by all scalar and vector
-  // types other than dvec3 or dvec4 - which require two slots each.
-
-  // Types: matrix
-  // The matrix component type and 'm' dimension value are considered as if
-  // they were a vector type. This value is then multiplied by the 'n'
-  // dimension value.
-
-  //TODO: according to the glsl 4.4 spec this slot size doubling isn't applied to vertex shader input. Test this!
-
-  language::type_dimension_t d1 = 1U;
-  language::type_dimension_t d2 = 1U;
-
-  const language::type type = vd.get_type();
-
-  const language::type_id id = type.get_id();
-  const language::type_dimensions& dimensions = type.get_dimensions();
-
-  if(dimensions.is_matrix())
-  {
-    d1 = dimensions.m();
-    d2 = dimensions.n();
-  }
-  else if(dimensions.is_vector())
-  {
-    d1 = (language::is_row_vector(dimensions) ? dimensions.n() : dimensions.m());
-  }
-
-  return _location_next + ((((id == language::id_double) && (d1 > 2U)) ? 2U : 1U) * d2);
-}
-
-// output::layout_manager definitions
-
-ns::output::layout_manager::layout_manager(detail::enum_flags<layout_flags> flags) :
-  _layout_in(flags.has_flag<layout_flags::flag_in>() ? layout_flags::flag_in : layout_flags::flag_none),
-  _layout_out(flags.has_flag<layout_flags::flag_out>() ? layout_flags::flag_out : layout_flags::flag_none),
-  _layout_uniform(flags.has_flag<layout_flags::flag_uniform>() ? layout_flags::flag_uniform : layout_flags::flag_none)
-{
-}
-
-ns::output::layout_manager::layout_manager(layout_manager&& manager) :
-  _layout_in(std::move(manager._layout_in)),
-  _layout_out(std::move(manager._layout_out)),
-  _layout_uniform(std::move(manager._layout_uniform))
-{
-}
-
-ns::output::layout_map& ns::output::layout_manager::get_layout_map(const syntax::variable_declaration& vd)
-{
-  switch(vd._qualifier)
-  {
-    case core::qualifier_storage::in:
-      return _layout_in;
-      break;
-    case core::qualifier_storage::out:
-      return _layout_out;
-      break;
-    case core::qualifier_storage::uniform:
-      return _layout_uniform;
-      break;
-    default:
-      throw std::exception();//TODO: exception type and message...
-  }
-}
-
-// non-member definitions
-
-ns::detail::enum_flags<ns::output::layout_flags> ns::operator|(ns::detail::enum_flags<output::layout_flags> lhs, ns::output::layout_flags rhs)
-{
-  return lhs |= rhs;
-}
-
-ns::detail::enum_flags<ns::output::layout_flags> ns::operator&(ns::detail::enum_flags<output::layout_flags> lhs, ns::output::layout_flags rhs)
-{
-  return lhs &= rhs;
 }
