+++ sltl_test/src/intrinsic_test.cpp
@@ -3,7 +3,7 @@
 #include "vector.h"
 #include "shader.h"
 
-#include "output/output.h"
+#include "output/glsl/output_glsl.h"
 
 
 namespace
@@ -11,7 +11,7 @@ namespace
   std::wstring to_string(const sltl::shader& shader)
   {
     // Prepend a newline character to exactly match the raw string literals
-    return L'\n' + shader.apply_action<sltl::output>(sltl::output_version::none, sltl::output_flags::flag_indent_space);
+    return L'\n' + shader.apply_action<sltl::glsl::output_glsl>(sltl::glsl::output_version::none, sltl::output_flags::flag_indent_space);
   }
 }
